requirements.txt
Django==4.2.3
djangorestframework==3.14.0
psycopg2-binary
gunicorn==20.1.0
python-telegram-bot==20.0